Refactor agent tool imports
